I decided to just go back to standard RCSV format for the records blocks because if there was a "" (an escaped intentional double-quote) in the description at the start of a line..that could get messed up!  We automatically messed it up though, so we can perfectly un-mess it up XD So that's what I did X3 (And updated the parser code correspondingly)
Sorry it's harder to read now though X'D
But it turns out pdfplumber is *so good at its job* that there were
absolutely no errors like with Tabula! :D    (which was the primary
reason to separate the binary format records with newlines in the first
place X3 )

So since pdfplumber is so good, we can just leave any errors in place in
future such projects, and rely on the parser to recognize WOAH IT'S SO
WRONG XD if there are any multiline-cell/row confusion errors since
(apparently!) those are going to be so extremely rare (if any at all!)
now! :D