Converted to the same format as Chapter 4 controls!  A very exact RCSV-based format.
I'm not sure if we want to exclusively use this in the future..  That
indentation-based format is *very* nice to look at and deal with
double-quotes in!  (no escape syntax!!)

But it was faster for me to do this than make a parser XD''
And if I ever do make a parser, I can make a serializer easy, and make a
format converter to automatically produce the indentation-based ("Basics
Process") format from the other one (and produce it for the Chapter 4
controls!) and this body of stuff here will serve as perfect testing
data for the format converter! :D